refactor code

+++ SportBooking.BLL/Dtos/ReservationResponseDto.cs
@@ -5,7 +5,7 @@ public class ReservationResponseDto
     public int Id { get; set; }
     public string SportFieldTitle { get; set; }
     public string ReservationTitle { get; set; }
-    public int TotalPrice { get; set; }
+    public double TotalPrice { get; set; }
     public string UserId { get; set; }
     public DateTime Start { get; set; }
     public DateTime End { get; set; }
